# swagger.json
{
    "openapi":"3.0.0",
    "info":{
        "version":"1.2.3",
        "description":"API for CloudKitchen application",
        "title":"CloudKitchen API"
    },
    "servers":[
{
    "url":"http://localhost:1400"
}    ],
"components":{
    "securitySchemes":{
                    "JWT":{
            "type":"apiKey",
            "name":"Authorization",
            "in":"header"
        }

        }
    },
    "paths":{
        "/api/menu":{
            "get":{
                "summary":"Get Menu",
                "description":"User will get full menu",
                "security":[
                    {"JWT":{
                        
                    }
                }
                ],
                "responses":{
                    "200":{
                        "description":"OK"
                    },
                    "401":{
                        "description":"Unauthorized"
                    }
                }
            }
        },
                "/api/cart":{
            "get":{
                "summary":"Get Cart Items",
                "description":"User will able to see items added on cart",
                "security":[
                    {"JWT":{
                        
                    }
                }
                ],
                "responses":{
                    "200":{
                        "description":"OK"
                    },
                    "401":{
                        "description":"Unauthorized"
                    }
                }
            }
        },

        "/api/users/signin":{
            "post":{
                "summary":"Login",
                "description":"User login to get token",
                "requestBody":{
                    "content":
                    {
                        "application/json":{
                        "schema":{
                            "type":"object",
                            "properties":{
                                "email":{
                                    "type":"string"
                                },
                                "password":{
                                    "type":"string"
                                }
                            }
                        }
                    }
                    }
                },
                "responses":{
                    "200":{
                        "description":"OK"
                    },
                    "400":{
                        "description":"Incorrect Credentials"
                    }
                }
            }
        }
    }
}